refactor: Separate getParameterList into seperate hook.
